+++ lib/view/weather.dart
